moved system console commands into their own folder
