+++ main/syscore/console.h
@@ -1,0 +1,6 @@
+#ifndef __JOLT_SYSCORE_CONSOLE_H__
+#define __JOLT_SYSCORE_CONSOLE_H__
+
+void console_syscore_register();
+
+#endif
