--- main/uart.c
@@ -1,107 +1,0 @@
-/* Jolt Wallet - Open Source Cryptocurrency Hardware Wallet
- Copyright (C) 2018  Brian Pugh, James Coxon, Michael Smaili
- https://www.joltwallet.com/
- 
- This program is free software; you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation; either version 3 of the License, or
- (at your option) any later version.
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
- You should have received a copy of the GNU General Public License
- along with this program; if not, write to the Free Software Foundation,
- Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301  USA
- */
-
-#include <string.h>
-#include <stdio.h>
-#include "esp_log.h"
-#include "freertos/FreeRTOS.h"
-#include "freertos/task.h"
-
-#include "uart.h"
-
-static const char TAG[] = "uart";
-
-
-void get_serial_input(char *serial_rx, const int buffersize){
-    
-    int line_pos = 0;
-    
-    while(1){
-        int c = getchar();
-        
-        if(c < 0) {
-            vTaskDelay(10 / portTICK_PERIOD_MS);
-            continue;
-        }
-        if(c == '\r') {
-            continue;
-        }
-        if(c == '\n') {
-            // terminate the string
-            serial_rx[line_pos] = '\0';
-            printf("\n");
-            break;
-        }
-        else {
-            putchar(c);
-            serial_rx[line_pos] = c;
-            line_pos++;
-            
-            // buffer full!
-            if(line_pos == buffersize) {
-                ESP_LOGI(TAG, "Command buffer full.");
-                serial_rx[line_pos] = '\0';
-                break;
-            }
-        }
-    }
-}
-
-void get_serial_input_int(char *serial_rx, const int buffersize){
-    // fills up serial_rx with an ascii string where all characters must be ints
-    
-    int line_pos = 0;
-    
-    while(1){
-        int c = getchar();
-        
-        if(c < 0) {
-            vTaskDelay(10 / portTICK_PERIOD_MS);
-            continue;
-        }
-        if(c == '\r') continue;
-        if(c == '\n') {
-            // terminate the string
-            serial_rx[line_pos] = '\0';
-            printf("\n");
-            break;
-        }
-        else {
-            if( c >= '0' && c <= '9' ){
-                putchar(c);
-                serial_rx[line_pos] = c;
-                line_pos++;
-                
-                // buffer full!
-                if(line_pos == buffersize) {
-                    ESP_LOGI(TAG, "Command buffer full.");
-                    serial_rx[line_pos] = '\0';
-                    break;
-                }
-            }
-        }
-    }
-}
-
-void flush_uart(){
-    //This is a terrible hack to flush the uarts buffer, a far better option would be rewrite all uart related code
-    // to use proper uart code from driver/uart.h
-    for(int bad_hack = 0; bad_hack <= 10; bad_hack++){
-        getchar();
-    };
-}
-
